+++ src/cofi/solvers/scipy_lstsq.py
@@ -1,5 +1,4 @@
 import inspect
-from re import L
 import numpy as np
 from scipy.linalg import lstsq
 
@@ -11,11 +10,11 @@ class ScipyLstSqSolver(BaseSolver):
 
     There are four cases:
 
-    1. basic case, where neither data noise nor regularisation is taken into account,
+    1. basic case, where neither data noise nor regularization is taken into account,
        :math:`m=(G^TG)^{-1}G^Td`
     2. with uncertainty, :math:`m=(G^TC_d^{-1}G)^{-1}G^TC_d^{-1}d`
-    3. with Tikhonov regularisation, :math:`m=(G^TG+\lambda L^TL)^{-1}G^Td`
-    4. with both uncertainty and regularisation,
+    3. with Tikhonov regularization, :math:`m=(G^TG+\lambda L^TL)^{-1}G^Td`
+    4. with both uncertainty and regularization,
        :math:`m=(G^TC_d^{-1}G+\lambda L^TL)^{-1}G^TC_d^{-1}d`
 
     where:
@@ -23,9 +22,9 @@ class ScipyLstSqSolver(BaseSolver):
     - :math:`m` refers to inferred model (solution)
     - :math:`G` refers to the Jacobian of the forward operator
     - :math:`d` refers to the data vector
-    - :math:`\lambda` refers to the regularisation factor that adjusts the ratio of
-      data misfit and regularisation term
-    - :math:`L` refers to the regularisation matrix, and is usually chosen to be
+    - :math:`\lambda` refers to the regularization factor that adjusts the ratio of
+      data misfit and regularization term
+    - :math:`L` refers to the regularization matrix, and is usually chosen to be
       damping (zeroth order Tikhonov), roughening (first order Tikhonov), or smoothing
       (second order Tikhonov).
     """
@@ -44,7 +43,7 @@ class ScipyLstSqSolver(BaseSolver):
     optional_in_problem: dict = {
         "data_covariance_inv": None,
         "data_covariance": None,
-        "regularisation_matrix": None,
+        "regularization_matrix": None,
     }
     required_in_options: set = {}
     optional_in_options: dict = {
@@ -99,8 +98,10 @@ class ScipyLstSqSolver(BaseSolver):
             is_diagonal = (self._Cd_inv == 0).all()
             np.fill_diagonal(self._Cd_inv, diag_elem)
             if is_diagonal:
+                print("is diagonal")
                 _gt_cdinv = self._G.T * diag_elem
             else:
+                print("not diagonal")
                 _gt_cdinv = self._G.T @ self._Cd_inv
             self._a = _gt_cdinv @ self._G
             self._b = _gt_cdinv @ self._d
@@ -108,24 +109,24 @@ class ScipyLstSqSolver(BaseSolver):
             self._a = self._G.T @ self._G
             self._b = self._G.T @ self._d
 
-        # check whether to take regularisation into account
+        # check whether to take regularization into account
         self._with_tikhonov = (
-            self._with_tikhonov_if_possible and inv_problem.regularisation_defined
+            self._with_tikhonov_if_possible and inv_problem.regularization_defined
         )
         # get lamda and L matrix if needed
         if self._with_tikhonov:
-            self._lamda = inv_problem.regularisation_factor
-            if inv_problem.regularisation_matrix_defined:
+            self._lamda = inv_problem.regularization_factor
+            if inv_problem.regularization_matrix_defined:
                 try:
-                    _L = inv_problem.regularisation_matrix(dummy_model)
-                    self._LtL = _L.T @ L
-                    self.components_used.append("regularisation_matrix")
+                    _L = inv_problem.regularization_matrix(dummy_model)
                 except Exception as exception:
                     raise ValueError(
-                        "regularisation matrix function isn't set properly for least "
+                        "regularization matrix function isn't set properly for least "
                         "squares solver, this should return a matrix unrelated to "
                         "model vector"
                     ) from exception
+                self._LtL = np.square(_L)
+                self.components_used.append("regularization_matrix")
             else:
                 self._LtL = np.identity(self._G.shape[1])
             self._a += self._lamda * self._LtL